+++ src/presenter/index.js
@@ -1,0 +1,1 @@
+export {default as TripPresenter} from "./trip-presenter";
