+++ src/presenter/trip-presenter.js
