+++ src/view/abstract-view.js
@@ -1,8 +1,8 @@
 import {createElement} from "../utils/render";
 
-export default class Abstract {
+export default class AbstractView {
   constructor() {
-    if (new.target === Abstract) {
+    if (new.target === AbstractView) {
       throw new Error(`Can't instantiate Abstract, only concrete one.`);
     }
