+++ src/view/days-view.js
@@ -1,4 +1,4 @@
-import AbstractView from "./abstract";
+import AbstractView from "./abstract-view";
 
 const createDaysTemplate = () => {
   return (
