+++ src/view/event-view.js
@@ -1,5 +1,5 @@
 import {EVENT_TYPE_ACTIVITY} from "../const";
-import AbstractView from "./abstract";
+import AbstractView from "./abstract-view";
 import {getTimeAtDefaultFormat, getDateAtSystemFormat, getDurationTime} from "../utils/date-formatters";
 
 const createAcceptedOffersTemplate = (offers) => {
