+++ src/view/index.js
@@ -1,10 +1,10 @@
-export {default as TripInfoView} from "./trip-info";
-export {default as NavigationControllerView} from "./nav-controller";
-export {default as EventFiltrationView} from "./event-filtration";
-export {default as TotalPriceView} from "./total-price";
-export {default as SortingView} from "./sorting";
-export {default as EventEditorView} from "./event-editor";
-export {default as DaysView} from "./days";
-export {default as DayView} from "./day";
-export {default as EventView} from "./event";
-export {default as NoEventView} from "./no-event";
+export {default as TripInfoView} from "./trip-info-view";
+export {default as NavigationControllerView} from "./nav-controller-view";
+export {default as EventFiltrationView} from "./event-filtration-view";
+export {default as TotalPriceView} from "./total-price-view";
+export {default as SortingView} from "./sorting-view";
+export {default as EventEditorView} from "./event-editor-view";
+export {default as DaysView} from "./days-view";
+export {default as DayView} from "./day-view";
+export {default as EventView} from "./event-view";
+export {default as NoEventView} from "./no-event-view";
